+++ dataprovider/foxsports/scraper_mlb_probable_starting_pitcher_test.go
@@ -10,7 +10,7 @@ import (
 	"github.com/stretchr/testify/assert"
 )
 
-func TestMLBProbableStartingPitcher(t *testing.T) {
+func TestMLBProbableStartingPitcherScraper(t *testing.T) {
 	if testing.Short() {
 		t.Skip("Skipping integration test")
 	}