+++ CHANGELOG.md
@@ -7,6 +7,15 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.12.0] - 2025-07-10
+### Added
+- Added `MatchupScraper` for baseball savant mlb (#84)
+- Added `FieldingBoxScoreScraper`,`BattingBoxScoreScraper`, `PitchingBoxScoreScraper`, and `PlayByPlayScraper` for baseball savant mlb (#86)
+### Changed
+- Moved `RFC3339ToTime` and `DateStrToTime` to time.go (#84)
+- Documentation updates (#84)
+- Deprecation warning on baseball reference feed (#87)
+
 ## [0.11.0] - 2025-07-02
 ### Added
 - Added `MatchupRunner`, `EventDataRunner`, `MatchupScraper`, and `EventDataScraper` abstractions to be reused across all data providers (#80)